backend: Make getPendingRequests() private
Change-type: patch
