+++ lib/core/backend/postgres/index.js
@@ -16,31 +16,10 @@ const locks = require('./locks')
 const cards = require('./cards')
 const streams = require('./streams')
 const queue = require('./queue')
+const utils = require('./utils')
 const markers = require('./markers')
 
 const SORT_KEY = '$$sort'
-
-const filter = (schema, element) => {
-	const result = skhema.filter(schema, element)
-
-	// If additionalProperties is false, remove properties that haven't been
-	// explicitly selected. This is done because the Skhema module will merge
-	// anyOf branches into the top level properties before applying the filter
-	// (which is correct) however, due to the Jellyfish permissions system,
-	// properties may be added to the result that the initial query did not
-	// specify, depending on which permission views are merged in.
-	if (schema.additionalProperties === false) {
-		for (const item of _.castArray(element)) {
-			for (const key in item) {
-				if (!schema.properties[key]) {
-					Reflect.deleteProperty(item, key)
-				}
-			}
-		}
-	}
-
-	return result
-}
 
 const defaultAdditionalPropertiesFalse = (schema) => {
 	// Default additionalProperties to false, to limit the ammount of retruned
@@ -96,7 +75,8 @@ const queryTable = async (context, backend, table, schema, options) => {
 				element.links = {}
 			}
 
-			element.links[linkType] = filter(linkSchema, element[`links.${linkSlug}`])
+			element.links[linkType] = utils.filter(
+				linkSchema, element[`links.${linkSlug}`])
 
 			Reflect.deleteProperty(element, `links.${linkSlug}`)
 		}
@@ -104,7 +84,7 @@ const queryTable = async (context, backend, table, schema, options) => {
 		return element
 	})
 
-	const result = filter(schema, elements)
+	const result = utils.filter(schema, elements)
 	const postProcessingEndDate = new Date()
 
 	logger.debug(context, 'Query database response', {
@@ -724,7 +704,7 @@ module.exports = class PostgresBackend {
 			schema.$$links &&
 			_.isNull(schema.$$links['is executed by']) &&
 			Object.keys(schema.$$links).length === 1) {
-			return this.getPendingRequests(context, {
+			return utils.getPendingRequests(context, this, {
 				limit: options.limit,
 				skip: options.skip,
 				mask: options.mask
@@ -820,26 +800,6 @@ module.exports = class PostgresBackend {
 	 */
 	async unlock (owner, slug) {
 		return locks.unlock(this.connection, this.errors, owner, slug)
-	}
-
-	/*
-	 * Get a set of pending action requests from the queue.
-	 *
-	 * Clients that require accessing the pending queue are advised
-	 * to use this function for performance reasons rather than doing
-	 * a more generic JSON Schema query on the cards table.
-	 */
-	async getPendingRequests (context, options = {}) {
-		const results = await queue.getElements(context, this.connection, {
-			table: 'requests',
-			limit: options.limit,
-			skip: options.skip
-		})
-
-		return filter(options.mask || {
-			type: 'object',
-			additionalProperties: true
-		}, results)
 	}
 
 	/*
